working on gpu

+++ CQL/d4rl/rlkit/torch/core.py
@@ -41,7 +41,7 @@ def _elem_or_tuple_to_variable(elem_or_tuple):
         return tuple(
             _elem_or_tuple_to_variable(e) for e in elem_or_tuple
         )
-    return ptu.from_numpy(elem_or_tuple)
+    return ptu.from_numpy(elem_or_tuple).float()
 
 
 def _filter_batch(np_batch):
